requirements.txt
pandas
h5py
supabase
python-dotenv
numpy
